refactor(editor): Extract contextmenu into features (no-changelog)
